Coverity: Fully initialize the |trace_| member variable.
Also changed the type of |count_| to size_t since it's a size variable.

CID_COUNT=1
CID=13504
BUG=none
TEST=none
R=brettw


Review URL: https://chromiumcodereview.appspot.com/10834006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148420 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/debug/stack_trace.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -19,7 +19,7 @@ StackTrace::StackTrace(const void* const* trace, size_t count) {
   count = std::min(count, arraysize(trace_));
   if (count)
     memcpy(trace_, trace, count * sizeof(trace_[0]));
-  count_ = static_cast<int>(count);
+  count_ = count;
 }
 
 StackTrace::~StackTrace() {
